track renamed experimental controllers and services
