+++ nodejs-server/controllers/ArExperimental.js
@@ -1,10 +1,10 @@
 'use strict';
 
 var utils = require('../utils/writer.js');
-var Ar = require('../service/ArService');
+var ArExperimental = require('../service/ArExperimentalService');
 
 module.exports.findAR = function findAR (req, res, next, date, _id) {
-  Ar.findAR(date, _id)
+  ArExperimental.findAR(date, _id)
     .then(function (response) {
       utils.writeJson(res, response);
     })
