+++ nodejs-server/controllers/CovarExperimental.js
@@ -1,10 +1,10 @@
 'use strict';
 
 var utils = require('../utils/writer.js');
-var Covar = require('../service/CovarService');
+var CovarExperimental = require('../service/CovarExperimentalService');
 
 module.exports.findCovar = function findCovar (req, res, next, lat, lon, forcastDays) {
-  Covar.findCovar(lat, lon, forcastDays)
+  CovarExperimental.findCovar(lat, lon, forcastDays)
     .then(function (response) {
       utils.writeJson(res, response);
     })
@@ -14,7 +14,7 @@ module.exports.findCovar = function findCovar (req, res, next, lat, lon, forcast
 };
 
 module.exports.sumCovar = function sumCovar (req, res, next, lat, lon, forcastDays, polygon) {
-  Covar.sumCovar(lat, lon, forcastDays, polygon)
+  CovarExperimental.sumCovar(lat, lon, forcastDays, polygon)
     .then(function (response) {
       utils.writeJson(res, response);
     })
